refactor : 전략 패턴 패키지 위치 변경
