+++ src/main/java/lotto/strategy/ExtractStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.strategy;
 
 import java.util.ArrayList;
 import java.util.List;
